全体的に検証用のプログラムをdemoディレクトリに移動
