+++ example-app/SantokuApp/android/app/src/main/java/jp/fintan/mobile/santokuapp/MainApplication.java
@@ -10,7 +10,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
 import jp.fintan.mobile.santokuapp.generated.BasePackageList;
-import jp.fintan.mobile.santokuapp.throwerror.ThrowErrorPackage;
+import jp.fintan.mobile.santokuapp.demo.throwerror.ThrowErrorPackage;
 
 import org.unimodules.adapters.react.ModuleRegistryAdapter;
 import org.unimodules.adapters.react.ReactModuleRegistryProvider;
