+++ example-app/SantokuApp/ios/Demo/RCTThrowErrorModule.m
