+++ example-app/SantokuApp/src/screens/demo/error/ErrorInEventHandler.tsx
