+++ example-app/SantokuApp/src/screens/demo/error/ErrorInReactComponent.tsx
